+++ internal/provider/provider.go
@@ -18,9 +18,10 @@ func New(version string) func() *schema.Provider {
 	return func() *schema.Provider {
 		p := &schema.Provider{
 			DataSourcesMap: map[string]*schema.Resource{
-				"utils_deep_merge_json":   dataSourceDeepMergeJSON(),
-				"utils_deep_merge_yaml":   dataSourceDeepMergeYAML(),
-				"utils_stack_config_yaml": dataSourceStackConfigYAML(),
+				"utils_deep_merge_json":        dataSourceDeepMergeJSON(),
+				"utils_deep_merge_yaml":        dataSourceDeepMergeYAML(),
+				"utils_stack_config_yaml":      dataSourceStackConfigYAML(),
+				"utils_spacelift_stack_config": dataSourceSpaceliftStackConfig(),
 			},
 		}