+++ internal/stack/stack_processor.go
@@ -9,51 +9,78 @@ import (
 	"io/ioutil"
 	"path"
 	"sort"
+	"strings"
+	"sync"
 )
 
 // ProcessYAMLConfigFiles takes a list of paths to YAML config files, processes and deep-merges all imports,
 // and returns a list of stack configs
-func ProcessYAMLConfigFiles(filePaths []string, processStackDeps bool, processComponentDeps bool) ([]string, error) {
-	var result []string
+func ProcessYAMLConfigFiles(filePaths []string, processStackDeps bool, processComponentDeps bool) ([]string, map[string]interface{}, error) {
+	count := len(filePaths)
+	listResult := make([]string, count)
+	mapResult := map[string]interface{}{}
+	var errorResult error
+	var wg sync.WaitGroup
+	wg.Add(count)
+	mu := &sync.Mutex{}
 
-	for _, p := range filePaths {
-		config, importsConfig, err := ProcessYAMLConfigFile(p, map[string]map[interface{}]interface{}{})
-		if err != nil {
-			return nil, err
-		}
+	for i, filePath := range filePaths {
+		go func(i int, p string) {
+			defer wg.Done()
 
-		var imports []string
-		for k := range importsConfig {
-			imports = append(imports, k)
-		}
-
-		uniqueImports := u.UniqueStrings(imports)
-		sort.Strings(uniqueImports)
-
-		componentStackMap := map[string]map[string][]string{}
-		if processStackDeps {
-			componentStackMap, err = createComponentStackMap(p)
+			config, importsConfig, err := ProcessYAMLConfigFile(p, map[string]map[interface{}]interface{}{})
 			if err != nil {
-				return nil, err
+				errorResult = err
+				return
 			}
-		}
 
-		finalConfig, err := ProcessConfig(p, config, processStackDeps, processComponentDeps, "", componentStackMap, importsConfig)
-		if err != nil {
-			return nil, err
-		}
+			var imports []string
+			for k := range importsConfig {
+				imports = append(imports, k)
+			}
 
-		finalConfig["imports"] = uniqueImports
+			uniqueImports := u.UniqueStrings(imports)
+			sort.Strings(uniqueImports)
 
-		yamlConfig, err := yaml.Marshal(finalConfig)
-		if err != nil {
-			return nil, err
-		}
+			componentStackMap := map[string]map[string][]string{}
+			if processStackDeps {
+				componentStackMap, err = CreateComponentStackMap(p)
+				if err != nil {
+					errorResult = err
+					return
+				}
+			}
 
-		result = append(result, string(yamlConfig))
+			finalConfig, err := ProcessConfig(p, config, processStackDeps, processComponentDeps, "", componentStackMap, importsConfig)
+			if err != nil {
+				errorResult = err
+				return
+			}
+
+			finalConfig["imports"] = uniqueImports
+
+			yamlConfig, err := yaml.Marshal(finalConfig)
+			if err != nil {
+				errorResult = err
+				return
+			}
+
+			stackName := strings.TrimSuffix(strings.TrimSuffix(path.Base(p), ".yaml"), ".yml")
+
+			mu.Lock()
+			defer mu.Unlock()
+
+			listResult[i] = string(yamlConfig)
+			mapResult[stackName] = finalConfig
+		}(i, filePath)
 	}
 
-	return result, nil
+	wg.Wait()
+
+	if errorResult != nil {
+		return nil, nil, errorResult
+	}
+	return listResult, mapResult, nil
 }
 
 // ProcessYAMLConfigFile takes a path to a YAML config file,
@@ -80,17 +107,36 @@ func ProcessYAMLConfigFile(
 	if importsSection, ok := stackMapConfig["import"]; ok {
 		imports := importsSection.([]interface{})
 
-		for _, i := range imports {
-			imp := i.(string)
+		count := len(imports)
+		var errorResult error
+		var wg sync.WaitGroup
+		wg.Add(count)
+		mu := &sync.Mutex{}
+
+		for _, im := range imports {
+			imp := im.(string)
 			p := path.Join(dir, imp+".yaml")
 
-			yamlConfig, _, err := ProcessYAMLConfigFile(p, importsConfig)
-			if err != nil {
-				return nil, nil, err
-			}
+			go func(p string) {
+				defer wg.Done()
 
-			configs = append(configs, yamlConfig)
-			importsConfig[imp] = yamlConfig
+				yamlConfig, _, err := ProcessYAMLConfigFile(p, importsConfig)
+				if err != nil {
+					errorResult = err
+					return
+				}
+
+				mu.Lock()
+				defer mu.Unlock()
+				configs = append(configs, yamlConfig)
+				importsConfig[imp] = yamlConfig
+			}(p)
+		}
+
+		wg.Wait()
+
+		if errorResult != nil {
+			return nil, nil, err
 		}
 	}
 
@@ -107,13 +153,16 @@ func ProcessYAMLConfigFile(
 
 // ProcessConfig takes a raw stack config, deep-merges all variables, settings, environments and backends,
 // and returns the final stack configuration for all Terraform and helmfile components
-func ProcessConfig(stack string,
+func ProcessConfig(
+	stack string,
 	config map[interface{}]interface{},
 	processStackDeps bool,
 	processComponentDeps bool,
 	componentTypeFilter string,
 	componentStackMap map[string]map[string][]string,
 	importsConfig map[string]map[interface{}]interface{}) (map[interface{}]interface{}, error) {
+
+	stackName := strings.TrimSuffix(strings.TrimSuffix(path.Base(stack), ".yaml"), ".yml")
 
 	globalVarsSection := map[interface{}]interface{}{}
 	globalSettingsSection := map[interface{}]interface{}{}
@@ -323,7 +372,7 @@ func ProcessConfig(stack string,
 				}
 
 				if processStackDeps == true {
-					componentStacks, err := findComponentStacks("terraform", component.(string), baseComponentName, componentStackMap)
+					componentStacks, err := FindComponentStacks("terraform", component.(string), baseComponentName, componentStackMap)
 					if err != nil {
 						return nil, err
 					}
@@ -333,7 +382,7 @@ func ProcessConfig(stack string,
 				}
 
 				if processComponentDeps == true {
-					componentDeps, err := findComponentDependencies("terraform", component.(string), baseComponentName, importsConfig)
+					componentDeps, err := FindComponentDependencies(stackName, "terraform", component.(string), baseComponentName, importsConfig)
 					if err != nil {
 						return nil, err
 					}
@@ -391,7 +440,7 @@ func ProcessConfig(stack string,
 				comp["env"] = finalComponentEnv
 
 				if processStackDeps == true {
-					componentStacks, err := findComponentStacks("helmfile", component.(string), "", componentStackMap)
+					componentStacks, err := FindComponentStacks("helmfile", component.(string), "", componentStackMap)
 					if err != nil {
 						return nil, err
 					}
@@ -401,7 +450,7 @@ func ProcessConfig(stack string,
 				}
 
 				if processComponentDeps == true {
-					componentDeps, err := findComponentDependencies("helmfile", component.(string), "", importsConfig)
+					componentDeps, err := FindComponentDependencies(stackName, "helmfile", component.(string), "", importsConfig)
 					if err != nil {
 						return nil, err
 					}