+++ internal/stack/stack_processor_test.go
@@ -2,6 +2,7 @@ package stack
 
 import (
 	c "github.com/cloudposse/terraform-provider-utils/internal/convert"
+	u "github.com/cloudposse/terraform-provider-utils/internal/utils"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -11,27 +12,41 @@ import (
 func TestStackProcessor(t *testing.T) {
 	filePaths := []string{
 		"../../examples/data-sources/utils_stack_config_yaml/stacks/uw2-dev.yaml",
+		"../../examples/data-sources/utils_stack_config_yaml/stacks/uw2-prod.yaml",
+		"../../examples/data-sources/utils_stack_config_yaml/stacks/uw2-staging.yaml",
+		"../../examples/data-sources/utils_stack_config_yaml/stacks/uw2-uat.yaml",
 	}
 
 	processStackDeps := true
 	processComponentDeps := true
 
-	var yamlResult, err = ProcessYAMLConfigFiles(filePaths, processStackDeps, processComponentDeps)
+	var listResult, mapResult, err = ProcessYAMLConfigFiles(filePaths, processStackDeps, processComponentDeps)
 	assert.Nil(t, err)
-	assert.Equal(t, len(yamlResult), 1)
+	assert.Equal(t, 4, len(listResult))
+	assert.Equal(t, 4, len(mapResult))
 
-	mapResult, err := c.YAMLToMapOfInterfaces(yamlResult[0])
+	mapResultKeys := u.StringKeysFromMap(mapResult)
+	assert.Equal(t, "uw2-dev", mapResultKeys[0])
+	assert.Equal(t, "uw2-prod", mapResultKeys[1])
+	assert.Equal(t, "uw2-staging", mapResultKeys[2])
+	assert.Equal(t, "uw2-uat", mapResultKeys[3])
+
+	mapConfig1, err := c.YAMLToMapOfInterfaces(listResult[0])
 	assert.Nil(t, err)
 
-	terraformComponents := mapResult["components"].(map[interface{}]interface{})["terraform"].(map[interface{}]interface{})
-	helmfileComponents := mapResult["components"].(map[interface{}]interface{})["helmfile"].(map[interface{}]interface{})
-	imports := mapResult["imports"].([]interface{})
+	terraformComponents := mapConfig1["components"].(map[interface{}]interface{})["terraform"].(map[interface{}]interface{})
+	helmfileComponents := mapConfig1["components"].(map[interface{}]interface{})["helmfile"].(map[interface{}]interface{})
+
+	imports := mapConfig1["imports"].([]interface{})
+
+	mapConfig2 := mapResult["uw2-dev"]
+	assert.Equal(t, len(imports), len(mapConfig2.(map[interface{}]interface{})["imports"].([]string)))
 
 	auroraPostgresComponent := terraformComponents["aurora-postgres"].(map[interface{}]interface{})
 	auroraPostgres2Component := terraformComponents["aurora-postgres-2"].(map[interface{}]interface{})
 
-	assert.Equal(t, auroraPostgres2Component["component"], "aurora-postgres")
-	assert.Equal(t, auroraPostgres2Component["settings"].(map[interface{}]interface{})["spacelift"].(map[interface{}]interface{})["branch"], "dev")
+	assert.Equal(t, "aurora-postgres", auroraPostgres2Component["component"])
+	assert.Equal(t, "dev", auroraPostgres2Component["settings"].(map[interface{}]interface{})["spacelift"].(map[interface{}]interface{})["branch"])
 	assert.Equal(t, "db.r4.xlarge", auroraPostgres2Component["vars"].(map[interface{}]interface{})["instance_type"])
 	assert.Equal(t, "test1_override2", auroraPostgres2Component["env"].(map[interface{}]interface{})["ENV_TEST_1"].(string))
 	assert.Equal(t, "test2_override2", auroraPostgres2Component["env"].(map[interface{}]interface{})["ENV_TEST_2"].(string))
@@ -56,11 +71,13 @@ func TestStackProcessor(t *testing.T) {
 	if processComponentDeps {
 		assert.Equal(t, "catalog/rds-defaults", auroraPostgresComponent["deps"].([]interface{})[0])
 		assert.Equal(t, "globals", auroraPostgresComponent["deps"].([]interface{})[1])
-		assert.Equal(t, "uw2-globals", auroraPostgresComponent["deps"].([]interface{})[2])
+		assert.Equal(t, "uw2-dev", auroraPostgresComponent["deps"].([]interface{})[2])
+		assert.Equal(t, "uw2-globals", auroraPostgresComponent["deps"].([]interface{})[3])
 
 		assert.Equal(t, "catalog/rds-defaults", auroraPostgres2Component["deps"].([]interface{})[0])
 		assert.Equal(t, "globals", auroraPostgres2Component["deps"].([]interface{})[1])
-		assert.Equal(t, "uw2-globals", auroraPostgres2Component["deps"].([]interface{})[2])
+		assert.Equal(t, "uw2-dev", auroraPostgres2Component["deps"].([]interface{})[2])
+		assert.Equal(t, "uw2-globals", auroraPostgres2Component["deps"].([]interface{})[3])
 	}
 
 	eksComponent := terraformComponents["eks"].(map[interface{}]interface{})
@@ -93,7 +110,7 @@ func TestStackProcessor(t *testing.T) {
 	assert.Equal(t, "globals", imports[3])
 	assert.Equal(t, "uw2-globals", imports[4])
 
-	yamlConfig, err := yaml.Marshal(mapResult)
+	yamlConfig, err := yaml.Marshal(mapConfig1)
 	assert.Nil(t, err)
 	t.Log(string(yamlConfig))
 }